Revert "Pull request for unified profile and fixed push and pop screens"
