+++ src/ducks/User.js
@@ -679,11 +679,7 @@ export const goToProfile = () => {
         dispatch({
             type: ACTIONS.GO_TO_PROFILE
         });
-        // Actions.pop();
-        // Actions.Profile()
-        
-        // Actions.pop()
-        Actions.push('profile')
+        Actions.profile();
     }
 };
