+++ src/screens/Leaderboard.js
@@ -5,7 +5,7 @@ import { NavBar, FilterPicker } from '../components/general'
 import {
   fetchMembersPoints,
   fetchMemberProfile,
-  goToProfile,
+  goToOtherProfile,
   pageLoad,
   getPrivilege,
   loadUser,
@@ -97,12 +97,8 @@ class Leaderboard extends Component {
     // }
 
     // if(item.points !== 0){
-
-  //Member List
-  //action(item.id)
       return (
         // <TouchableOpacity onPress = {() => action(item.id)}>
-        <TouchableOpacity onPress = {() => this.callUser(item.id)}>
           <View style={contentContainerStyle}>
               <View style={containerStyle}>
               <Image   
@@ -131,7 +127,7 @@ class Leaderboard extends Component {
               
               </View>
           </View>
-          </TouchableOpacity>
+        // </TouchableOpacity>
       )
     }
 
@@ -142,9 +138,7 @@ class Leaderboard extends Component {
   callUser(id){
     this.props.pageLoad();
     this.props.fetchMemberProfile(id);
-    //this.props.goToOtherProfile();
-    Actions.pop();
-    Actions.profile();
+    this.props.goToOtherProfile();
   }
   
 }
@@ -215,7 +209,7 @@ const mapStateToProps = ({ user, members, general }) => {
 const mapDispatchToProps = {
   fetchMembersPoints,
   fetchMemberProfile,
-  goToProfile,
+  goToOtherProfile,
   pageLoad,
   getPrivilege,
   loadUser,
