+++ src/screens/OtherProfile.js
@@ -2,211 +2,257 @@ import React, { Component } from 'react';
 import { Actions } from 'react-native-router-flux';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import { connect } from 'react-redux';
-import { Button, Spinner, NavBar } from '../components/general'
-import { loadUser, logoutUser, goToEditOtherProfileForm, pageLoad} from '../ducks';
-import { Text, View, StyleSheet, Image, ScrollView, TouchableOpacity, Dimensions } from 'react-native';
+import {Button, Spinner} from '../components/general'
+import { goToEditOtherProfileForm } from '../ducks';
+import {
+  Text,
+  View, StyleSheet,
+  Image,
+  ScrollView,
+  TouchableOpacity,
+ 	Dimensions
+	} from 'react-native';
 import { Avatar, Divider } from 'react-native-elements';
 
 const dimension = Dimensions.get('window');
+
 class OtherProfile extends Component {
 
   render() {
-    return (
-          this.renderContent()
-      )
+    // alert(this.props.loading)
+     if(this.props.loading){
+      return <Spinner>{this.renderContent}</Spinner>
     }
-
+    else return (
+      <View>
+        {this.renderContent()}
+      </View>
+    )
+  }
+  
   renderContent(){
-    const { firstName, lastName, email, major, points, quote } = this.props;
+    const { firstName, lastName, email, major, points, picture, quote } = this.props;
 
     const {
-      bioContainer,
-      taglineContainer,
-      fieldContainerStyle,
-      itemLabelText,
-      itemValueText,
-      textColor
-      } = styles
-
-    return (
-      <View style={{flex: 1}}>
-        <Text style={[textColor,]}>Email:</Text>
-        <NavBar title="Profile" back onBack={() => Actions.pop()} />
-        {this.renderPicture()}
-        
-          <View style={bioContainer}>
-            <View style={taglineContainer}>
-                <Text style={[itemLabelText, textColor, {flex: 1}]}>{firstName + ' ' + lastName}</Text>
-            </View>
-            <View style={fieldContainerStyle}>
-              <Text style={[itemLabelText, textColor]}>Email:</Text>
-              <Text style={[itemValueText, textColor]}>{email}</Text>
-            </View>
-            <View style={fieldContainerStyle}>
-              <Text style={[itemLabelText, textColor]}>Major:</Text>
-              <Text style={[itemValueText, textColor]}>{major}</Text>
-            </View>
-            <TouchableOpacity style = {fieldContainerStyle} onPress={() => {
-              Actions.pointsBreakDown()}}
-            >
-              <Text style={[itemLabelText, textColor]}>Points:</Text>
-              <Text style={[itemValueText, textColor]}>{points}</Text>
-            </TouchableOpacity>
-          </View>
-          {this.renderSocialMedia()}
-          {this.renderButtons()}
-      </View>
-  )
-
-  }
-
-  renderPicture() {
-    const {
+      bottomHalfContainerStyle,
+      containerStyle,
       headerInfoContainer,
-    } = styles
+      avatarContainerStyle,
+      taglineContainer,
+      taglineTextStyle,
+      contentContainerStyle,
+      contentItemsContainerStyle,
+      itemLabelContainerStyle,
+      itemLabelText,
+      itemValueContainerStyle,
+      itemValueText,
+      buttonsContainerStyle,
+      editButtonContainer,
+			editLogoContainer,
+      logOutButtonContainer } = styles,
+			dimension = Dimensions.get('window');
 
-    const {
-      firstName,
-      lastName,
-      picture
-    } = this.props
-
-    return (
-      <View style={headerInfoContainer}>
-        <Avatar
-          size="xlarge"
-          rounded
-          source={{uri: picture}}
-          title={`${firstName[0]}${lastName[0]}`}
-          onPress={() => alert("Coming Soon") }
-          activeOpacity={0.7}
-          />
-      </View>
+      return (
+        <ScrollView>
+          <View style={headerInfoContainer}>
+            <View style={avatarContainerStyle}>
+              <Avatar
+                large
+                rounded
+                source={{uri: picture}}
+                title={`${firstName[0]}${lastName[0]}`}
+                onPress={() => alert("Coming Soon") }
+                activeOpacity={0.7}
+                />
+            </View>
+            <View style={taglineContainer}>
+               <Text
+							 		style={{
+										color: 'white',
+										fontSize: 20,
+										fontWeight: 'bold',
+										textAlign: 'center'}}
+										>{firstName + ' ' + lastName}</Text>
+               <Text
+							 		style={{
+										color: 'white',
+										fontSize: 16,
+										textAlign: 'center',
+										lineHeight: 25,
+										width: dimension.width *.9}}
+										>{quote}</Text>
+            </View>
+              {this.renderSocialMedia()}
+          </View>
+         <View style={{backgroundColor: '#0c0b0b'}}>
+          <View style={contentContainerStyle}>
+            <View style={contentItemsContainerStyle}>
+              <View style={{ height: dimension.height *.1, backgroundColor: '#0c0b0b'}}>
+                <Text style={itemLabelText}></Text>
+              </View>
+              <View style={itemValueContainerStyle}>
+                <Text style={itemValueText}>{/*firstName + ' ' + lastName*/}</Text>
+              </View>
+            </View>
+            <View style={contentItemsContainerStyle}>
+              <View style={itemLabelContainerStyle}>
+                <Text style={itemLabelText}>Email:</Text>
+              </View>
+              <View style={itemValueContainerStyle}>
+                <Text style={itemValueText}>{email}</Text>
+              </View>
+            </View>
+            <View style={contentItemsContainerStyle}>
+              <View style={itemLabelContainerStyle}>
+                <Text style={itemLabelText}>Major:</Text>
+              </View>
+              <View style={itemValueContainerStyle}>
+                <Text style={itemValueText}>{major}</Text>
+              </View>
+            </View>
+            <View style={contentItemsContainerStyle}>
+              <View style={itemLabelContainerStyle}>
+                <Text style={itemLabelText}>Points:</Text>
+              </View>
+              <View style={itemValueContainerStyle}>
+                <Text style={itemValueText}>{points}</Text>
+              </View>
+            </View>
+          </View>
+					<View style={buttonsContainerStyle}>
+              {this.renderButtons()}
+          </View>
+        </View>
+        </ScrollView>
     )
+
   }
 
-
-  renderButtons(){
-    const {
-      buttonsContainerStyle
-    } = styles
-    return (
-      <View style={buttonsContainerStyle}>
-          <Button
-            title = "EDIT PROFILE"
-            onPress={this.props.goToEditOtherProfileForm.bind(this)}
-          />
-          <Button
-            title = "LOG OUT"
-            onPress={this.props.logoutUser.bind(this)}
-          />
-      </View>
-    )
+  renderEditButton(privilege){
+    if (privilege.board)
+      return (
+          <View>
+            <Button
+              title = "EDIT PROFILE"
+              onPress = {() => goToEditOtherProfileForm()}
+            />
+          </View>
+      )
   }
 
-    renderSocialMedia(){
+  renderButtons() {
     const {
-      LogoContainer,
-      socialmediarow
-    } = styles
+      privilege,
+      goToEditOtherProfileForm,
+      } = this.props;
+
     return (
-      <View style={socialmediarow}>
-			<Text style={{fontSize: 17, alignSelf:"center"}}> Social Media</Text>
-			<View style={{flexDirection: 'row'}}>
-        <View style= {LogoContainer}>
+      <View>
+      {this.renderEditButton(privilege)}
+        <View>
+          <Button
+            title = "BACK TO LEADERBOARD"
+            onPress={() => Actions.popTo("Leaderboard")}
+          />
+        </View>
+      </View>
+    )
+
+  }
+
+  renderSocialMedia(){
+    return (
+			<View style={styles.editLogoContainer}>
+        <View style= {styles.editLogoContainer}>
           <TouchableOpacity
-            onPress={() => {
-              alert("Coming Soon")
-              // Actions.PostShow({ title: 'Linkedin', uri: 'https://www.linkedin.com/'})
-            }
-            }>
-            <Ionicons name="logo-linkedin" size={40} color='#000000'/>
+            onPress={() => Actions.PostShow({ title: 'Linkedin', uri: 'https://www.linkedin.com/'})}>
+            <Ionicons name="logo-linkedin" size={40} color='#fff'/>
           </TouchableOpacity>
         </View>
-        <View style={LogoContainer}>
+        <View style={styles.editLogoContainer}>
           <TouchableOpacity
-            onPress={() => {
-              alert("Coming Soon")
-            // Actions.PostShow({ title: 'Github', uri: 'https://www.github.com/'})
-            }
-          }>
-            <Ionicons name="logo-github" size={40} color='#000000'/>
+            onPress={() => Actions.PostShow({ title: 'Github', uri: 'https://www.github.com/'})}>
+            <Ionicons name="logo-github" size={40} color='#fff'/>
           </TouchableOpacity>
         </View>
       </View>
-			</View>
     )
   }
 }
 
 const styles = StyleSheet.create({
-  headerInfoContainer: {
-    flex: .6,
-    backgroundColor: '#2C3239',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderBottomColor: '#e0e6ed22',
-    borderBottomWidth: 1,
-    padding: 15,
+  bottomHalfContainerStyle: {
+    backgroundColor: 'rgb(240,240,240)'
   },
-  textColor: {
-    color: '#e0e6ed'
+  headerInfoContainer: {
+    flex: 1,
+    paddingTop: 30,
+    paddingBottom: 50,
+    backgroundColor: '#2C3239'
+  },
+  avatarContainerStyle: {
+    justifyContent: 'center',
+    alignItems: 'center',
   },
   taglineContainer: {
-    flex: .2,
     alignItems: 'center',
-		marginTop: dimension.height * .02,
+    marginTop: 10,
   },
-  fieldContainerStyle: {
-    flex: .2,
+  taglineTextStyle:{
+    fontSize: 16,
+    fontWeight: '600'
+  },
+  contentContainerStyle: {
+		height: dimension.width,
+    flex: 1,
+    marginTop: 20,
+    marginBottom: 20,
+
+  },
+  contentItemsContainerStyle: {
     flexDirection: 'row',
+    paddingHorizontal: 15,
+    paddingVertical: 15,
+  },
+  itemLabelContainerStyle: {
+    flex: 1,
+    justifyContent: 'center'
   },
   itemLabelText: {
-    flex: .25,
     fontSize: 18,
     fontWeight: 'bold',
 		color: '#fff',
-		lineHeight: dimension.height * .03
+		lineHeight: 26
   },
   itemValueContainerStyle: {
     flex: 4,
-    flexDirection:'row',
     justifyContent: 'center',
     alignItems: 'flex-start'
   },
   itemValueText: {
-    flex: 1,
     fontSize: 16,
     fontWeight: '500',
 		color: '#fff',
+		lineHeight: 26
   },
   buttonsContainerStyle: {
-    flex: .4,
-    marginRight: dimension.height * .015,
-    marginLeft: dimension.height * .015,
-    height: dimension.height * .145,
+    marginRight: 10,
+    marginLeft: 10,
   },
-	LogoContainer: {
+  editButtonContainer: {
+    marginTop: 10,
+    marginBottom: 10,
+  },
+	editLogoContainer: {
     flex: 1,
-    marginTop: dimension.height * .002,
-    alignItems: 'center',
+		marginTop: 3,
+		flexDirection: 'row',
 		justifyContent: 'center'
 	},
-	socialmediarow: {
-    flex: .25,
-		paddingTop: dimension.height * .015,
-		paddingBottom: dimension.height * .015,
-		backgroundColor: '#dee0e2',
-	},
-	bioContainer: {
-    flex: 1,
-    backgroundColor: '#2C3239',
-    paddingLeft: '5%'
-	},
-	socialmediatext: {
-		flex:1,
-    alignSelf: 'center'
-	},
+  logOutButtonContainer: {
+    marginTop: 10,
+    marginBottom: 60
+  },
 });
 
 const mapStateToProps = ({ members, general, user }) => {
@@ -214,14 +260,11 @@ const mapStateToProps = ({ members, general, user }) => {
   const { loading } = general;
   const { privilege } = user;
 
-  return { firstName, lastName, email, major, points, picture, quote, loading };
+  return { firstName, lastName, email, major, points, picture, quote, loading, privilege };
 };
 
 const mapDispatchToProps = {
-  loadUser,
-  logoutUser,
-  goToEditOtherProfileForm,
-  pageLoad
+  goToEditOtherProfileForm
  };
 
 export default connect(mapStateToProps, mapDispatchToProps)(OtherProfile);
